+++ client/src/game/shapes/variants/asset.ts
@@ -152,7 +152,8 @@ export class Asset extends BaseRect implements IAsset {
         super.drawPost(ctx, lightRevealRender);
     }
 
-    setImage(url: string, sync: boolean): void {
+    setImage(assetId: AssetId, url: string, sync: boolean): void {
+        this.assetId = assetId;
         this.#loaded = false;
         this.src = url;
         this.img.src = baseAdjust(url);
@@ -160,6 +161,6 @@ export class Asset extends BaseRect implements IAsset {
             this.setLoaded();
         };
         const uuid = getGlobalId(this.id);
-        if (uuid && sync) sendAssetRectImageChange({ uuid, src: url });
+        if (uuid && sync) sendAssetRectImageChange({ uuid, src: url, assetId });
     }
 }