+++ server/src/api/models/shape/__init__.py
@@ -98,6 +98,7 @@ class ShapeTextSizeUpdate(TypeIdModel):
 class ShapeAssetImageSet(TypeIdModel):
     uuid: str = Field(json_schema_extra={"typeId": "GlobalId"})
     src: str
+    assetId: int = Field(json_schema_extra={"typeId": "AssetId"})
 
 
 class ShapeTemplateAdd(TypeIdModel):