+++ CHANGELOG.md
@@ -10,6 +10,11 @@ tech changes will usually be stripped from release notes for the public
 
 ## Unreleased
 
+### Fixed
+
+-   Asset pick modal butons don't work when clicking on their bottom half
+-   Asset changing not updating the assetId stored on the shape
+
 ## [2026.1]
 
 ### Added